Title: 낮은 지점들, Time: 566ms, Memory: 29MB, Status: Passed - CodeTree

+++ 240401/낮은 지점들/lowest-points.java
@@ -21,7 +21,7 @@ public class Main {
         h.put(x, y);
     }
 
-    int ans = 0;
+    long ans = 0;
     for (Integer i : h.values())
       ans += i;
